Consolidate extraction of tokens from annotated string field value.
