+++ searchlib/src/vespa/searchlib/util/linguisticsannotation.cpp
