+++ searchsummary/src/vespa/searchsummary/docsummary/CMakeLists.txt
@@ -23,7 +23,6 @@ vespa_add_library(searchsummary_docsummary OBJECT
     juniper_dfw_term_visitor.cpp
     juniper_query_adapter.cpp
     juniperproperties.cpp
-    linguisticsannotation.cpp
     matched_elements_filter_dfw.cpp
     positionsdfw.cpp
     query_term_filter.cpp
